The implementation of Legion was simplified, and an X botob was added
