--- src/Components/Moks/tweets.jsx
@@ -1,72 +1,0 @@
-export const tweets = [
-  {
-    id: "101",
-    avatar: "https://randomuser.me/api/portraits/men/45.jpg",
-    name: "John Doe",
-    handle: "@john_doe",
-    text: "Exploring new cities is my favorite way to spend the weekends! 🏙️✈️ #Travel",
-  },
-  {
-    id: "102",
-    avatar: "https://randomuser.me/api/portraits/women/35.jpg",
-    name: "Sophia Walker",
-    handle: "@sophia_w",
-    text: "Just baked the most delicious chocolate chip cookies 🍪❤️ #BakingLove #SweetTreats",
-  },
-  {
-    id: "103",
-    avatar: "https://randomuser.me/api/portraits/men/22.jpg",
-    name: "Michael Tech",
-    handle: "@tech_mike",
-    text: "The future of AI is beyond imagination. What do you think the next big thing will be? 🤖 #AI #Innovation",
-  },
-  {
-    id: "104",
-    avatar: "https://randomuser.me/api/portraits/women/44.jpg",
-    name: "Emily Green",
-    handle: "@emily_green",
-    text: "Feeling blessed after a peaceful walk in the park today 🌳☀️ #NatureLover",
-  },
-  {
-    id: "105",
-    avatar: "https://randomuser.me/api/portraits/men/33.jpg",
-    name: "Liam Fitness",
-    handle: "@fitliam",
-    text: "Crushed my workout this morning! Time to refuel with some protein 💪🥑 #FitnessGoals #HealthyLiving",
-  },
-  {
-    id: "106",
-    avatar: "https://randomuser.me/api/portraits/women/55.jpg",
-    name: "Ava Baker",
-    handle: "@ava_bakes",
-    text: "Today I tried a new cupcake recipe and they turned out amazing! 🧁😍 #BakingGoals",
-  },
-  {
-    id: "107",
-    avatar: "https://randomuser.me/api/portraits/men/11.jpg",
-    name: "Daniel Movies",
-    handle: "@moviedan",
-    text: "Can't stop thinking about the latest sci-fi movie. Absolutely mind-blowing! 🚀🎬 #SciFi",
-  },
-  {
-    id: "108",
-    avatar: "https://randomuser.me/api/portraits/women/21.jpg",
-    name: "Mia Music",
-    handle: "@miamusician",
-    text: "Just wrote a new song! So excited to share it with the world 🎶🎸 #Songwriting #NewMusic",
-  },
-  {
-    id: "109",
-    avatar: "https://randomuser.me/api/portraits/men/58.jpg",
-    name: "Chris Football",
-    handle: "@chris_football",
-    text: "What a game last night! That last-minute goal was incredible! ⚽️🔥 #FootballFan",
-  },
-  {
-    id: "110",
-    avatar: "https://randomuser.me/api/portraits/women/29.jpg",
-    name: "Ella Reads",
-    handle: "@bookella",
-    text: "Just finished reading an amazing book. Highly recommend it to everyone 📚✨ #BookLover #Reading",
-  },
-];
